backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML
anthropic>=0.40.0
google-cloud-vision==3.4.5

# Image Processing
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1